c version

+++ Pathfinder/inc/graph.h
@@ -6,10 +6,10 @@
 #include "libmx.h"
 
 typedef struct s_info {
-    t_vector graph;     // vector<vector<pair<ull, ull>>>
-    t_vector parents;   // vector<vector<ull>>
-    t_vector routes;    // vector<vector<ull>>
-    t_vector distances; // vector<ull>
+    t_vector graph;     // vector<vector<pair<size_t, size_t>>>
+    t_vector parents;   // vector<vector<size_t>>
+    t_vector routes;    // vector<vector<size_t>>
+    t_vector distances; // vector<size_t>
     t_vector visited;   // vector<char>
     t_vector names;     // vector<string>
     size_t start;       // entry point to graph
@@ -17,10 +17,8 @@ typedef struct s_info {
 } t_info;
 
 typedef struct s_pair {
-    unsigned long long island;
-    unsigned long long distance;
+    size_t island;
+    size_t distance;
 } t_pair;
-
-// t_info info = {{}, {}, {}, {}, {}, {}, 0, size};
 
 #endif
