+++ Pathfinder/libmx/src/file/mx_get_file_length.c
@@ -1,8 +1,8 @@
 #include "libmx.h"
 
-int mx_get_file_length(const char *filename) {
+long long mx_get_file_length(const char *filename) {
     int stream = open(filename, 0);
-    int length = 0;
+    long long length = 0;
 
     for (char buffer; read(stream, &buffer, 1) > 0; ++length);
