+++ Pathfinder/libmx/src/string/mx_strlen.c
@@ -1,7 +1,7 @@
 #include "libmx.h"
 
-int mx_strlen(const char *s) {
-    int len = 0;
+size_t mx_strlen(const char *s) {
+    size_t len = 0;
 
     for (; s[len]; ++len);
     return len;
